+++ rector.php
@@ -2,17 +2,34 @@
 
 use Rector\CodeQuality\Rector\Concat\JoinStringConcatRector;
 use Rector\Config\RectorConfig;
+use Rector\PHPUnit\CodeQuality\Rector\Class_\PreferPHPUnitSelfCallRector;
+use Rector\PHPUnit\Set\PHPUnitSetList;
 use Rector\Set\ValueObject\SetList;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->import(SetList::CODE_QUALITY);
-    $rectorConfig->import(SetList::PHP_71);
-    $rectorConfig->import(SetList::PHP_72);
-    $rectorConfig->import(SetList::PHP_73);
-    $rectorConfig->import(SetList::PHP_74);
-    $rectorConfig->import(SetList::PHP_80);
-    $rectorConfig->import(SetList::PHP_81);
-
+    $rectorConfig->sets([
+        SetList::CODE_QUALITY,
+        SetList::PHP_71,
+        SetList::PHP_72,
+        SetList::PHP_73,
+        SetList::PHP_74,
+        SetList::PHP_80,
+        SetList::PHP_81,
+        SetList::PHP_82,
+        PHPUnitSetList::PHPUNIT_40,
+        PHPUnitSetList::PHPUNIT_50,
+        PHPUnitSetList::PHPUNIT_60,
+        PHPUnitSetList::PHPUNIT_70,
+        PHPUnitSetList::PHPUNIT_80,
+        PHPUnitSetList::PHPUNIT_90,
+        PHPUnitSetList::PHPUNIT_100,
+        PHPUnitSetList::PHPUNIT_110,
+        PHPUnitSetList::PHPUNIT_CODE_QUALITY,
+        PHPUnitSetList::ANNOTATIONS_TO_ATTRIBUTES,
+    ]);
+    $rectorConfig->importNames();
+    $rectorConfig->importShortClasses(false);
+    $rectorConfig->rule(PreferPHPUnitSelfCallRector::class);
     $rectorConfig->skip([
         // skip csv test file to keep `\r` and `\n` for readability
         JoinStringConcatRector::class => [
@@ -20,10 +37,16 @@ return static function (RectorConfig $rectorConfig): void {
             __DIR__ . '/tests/Unit/CSVArrayTest.php',
         ],
     ]);
-
-    // paths to refactor; solid alternative to CLI arguments
-    $rectorConfig->paths([__DIR__ . '/src', __DIR__ . '/tests']);
-
-    // Path to PHPStan with extensions, that PHPStan in Rector uses to determine types
-    $rectorConfig->phpstanConfig(__DIR__ . '/phpstan.neon');
+    $rectorConfig->paths([
+        __DIR__ . '/src',
+        __DIR__ . '/tests',
+    ]);
+    $rectorConfig->bootstrapFiles([
+        // Rector uses PHPStan internally, which in turn requires Larastan to be set up correctly
+        __DIR__ . '/vendor/larastan/larastan/bootstrap.php',
+    ]);
+    $rectorConfig->phpstanConfigs([
+        __DIR__ . '/phpstan.neon',
+        __DIR__ . '/vendor/larastan/larastan/extension.neon',
+    ]);
 };