+++ src/Queue/DispatchJob.php
@@ -13,10 +13,6 @@ final class DispatchJob extends Command
     public function handle(): void
     {
         $class = $this->argument('class');
-        if (! is_string($class)) {
-            $notString = gettype($class);
-            throw new \Exception("Expected argument 'class' to be string, got {$notString}.");
-        }
         if (! class_exists($class)) {
             throw new \Exception("Job class {$class} does not exist.");
         }