+++ tests/TestCase.php
@@ -2,6 +2,8 @@
 
 namespace MLL\LaravelUtils\Tests;
 
+use Illuminate\Foundation\Application;
+use Illuminate\Support\ServiceProvider;
 use MLL\LaravelUtils\LaravelUtilsServiceProvider;
 use MLL\LaravelUtils\Mail\MailServiceProvider;
 use Orchestra\Testbench\TestCase as BaseTestCase;
@@ -11,12 +13,12 @@ abstract class TestCase extends BaseTestCase
     /**
      * Set when not in setUp.
      *
-     * @var \Illuminate\Foundation\Application
+     * @var Application
      */
     // @phpstan-ignore-next-line
     protected $app;
 
-    /** @return array<int, class-string<\Illuminate\Support\ServiceProvider>> */
+    /** @return array<int, class-string<ServiceProvider>> */
     protected function getPackageProviders($app): array
     {
         return [